requirements.txt
alembic==1.11.2
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cssselect==1.2.0
dateparser==1.1.8
demjson3==3.0.6
facebook-scraper==0.2.59
fake-useragent==1.2.1
Flask==2.3.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
google==3.0.0
googlesearch-python==1.2.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
parse==1.19.1
pyee==8.2.2
PyMySQL==1.1.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tqdm==4.66.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
w3lib==2.1.2
websockets==10.4
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.2
